pyproject.toml
[project]
name = "consulting-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "google-adk>=1.22.1",
    "litellm>=1.81.0",
    "python-dotenv>=1.2.1",
]
